+++ src/features/main/hooks/useGetGoods.ts
@@ -7,7 +7,7 @@ import {
   GetGoodsRequestParams,
   GetGoodsResponse,
 } from "@/features/main/api/getGoods";
-import { goodsInfiniteQueryOptions } from "@/features/main/queries/goodsInfiniteQueryOptions";
+import { goodsInfiniteQueryOptions } from "@/features/main/queries/goods";
 
 export const useGetGoods = (
   order: GetGoodsRequestParams["order"] = "recommended"
