+++ src/features/main/queries/goods.ts
